+++ event-sourcing/event-repository/event-repository-jdbc/src/main/java/uk/gov/justice/services/eventsourcing/repository/jdbc/PublishQueue.java
@@ -1,0 +1,28 @@
+package uk.gov.justice.services.eventsourcing.repository.jdbc;
+
+import java.time.ZonedDateTime;
+import java.util.Optional;
+import java.util.UUID;
+
+public interface PublishQueue {
+    enum PublishQueueTableName {
+
+
+        PRE_PUBLISH_QUEUE_TABLE("pre_publish_queue"),
+        PUBLISH_QUEUE_TABLE("publish_queue");
+
+        private final String tableName;
+
+        PublishQueueTableName(final String tableName) {
+            this.tableName = tableName;
+        }
+
+        public String getTableName() {
+            return tableName;
+        }
+    }
+
+    void addToQueue(final UUID eventId, final ZonedDateTime queuedAt);
+    Optional<UUID> popNextEventId();
+    int getSizeOfQueue();
+}